SRE-1942 Backport delete_on_termination attribute
We can't upgrade to a newer AWS cookbook because it depends on a newer
version of the 'ohai' cookbook than we're willing to upgrade to right
now (because our 'nginx' cookbook requires an old 'ohai'), but we want
to be able to mark EBS volumes to terminate with their associated
instances. This backports the ':delete_on_termination' attribute from
v4.1.1 of the AWS cookbook to our v2.7.2 fork.